python/requirements.txt
openai==0.28.0
pinecone-client==2.2.1
python-dotenv==1.0.0